Mudança no tipo do id de users

+++ database/migrations/2023_10_06_003747_create_produto.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->date('validade');
             $table->timestamps();
 
-            $table->unsignedSmallInteger('id_users_fk');
+            $table->unsignedBigInteger('id_users_fk');
             $table->foreign('id_users_fk')->references('id')->on('users');
         });
     }
