+++ database/migrations/2023_10_06_010906_create_estoque.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->boolean('status')->default(1);
             $table->timestamps();
 
-            $table->unsignedSmallInteger('id_users_fk');
+            $table->unsignedBigInteger('id_users_fk');
             $table->foreign('id_users_fk')->references('id')->on('users');
 
             $table->unsignedSmallInteger('id_marca_fk');
